# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Hybrid Swagger Documentation",
        "version": "1.0.0",
        "title": "Hybrid App API",
        "contact": {
            "email": "emachalan@agilesoftlabs.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "schemes": ["http"],
    
    "host": "localhost:8000",
    "basePath": "/",
    "paths" : {
        "/notes" : {
            "get" : {
                "summary" : "Get all notes",
                "description": "Get all notes",
                "tags": ["Note"],
                "produces": ["application/json"],
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "description": "note title to get",
                        "required": false,
                        "type": "string"
                        }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/NoteResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            },
            "post" : {
                "summary" : "Create Note",
                "description": "Create Note",
                "tags": ["Note"],
                "produces": ["application/json"],
                "consumes": ["application/json"],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "note object",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Note"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/NoteResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
        "/notes/{id}" : {
            "get" : {
                "summary" : "Get note by id",
                "description": "Get note by id",
                "tags": ["Note"],
                "produces": ["application/json"],
                "parameters": [
                    {
                    "name": "id",
                    "in": "path",
                    "description": "note id to get",
                    "required": true,
                    "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "object",
                            "items": {
                                "$ref": "#/definitions/NoteResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            },
            "put" : {
                "summary" : "Update note by id",
                "description": "Update note by id",
                "tags": ["Note"],
                "produces": ["application/json"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "note id that needs to be updated",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "note object",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Note"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "object",
                            "items": {
                                "$ref": "#/definitions/NoteResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        }
    }, 
    "definitions": {
        "NoteResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "Note": {
            "type": "object",
            "properties": {
                "title": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                }
            }
        },
        "InvalidResponse": {
            "type": "object",
            "properties": {
                "statusCode": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }

        }
    }
}